+++ src/main/java/hous/server/config/interceptor/auth/Auth.java
@@ -1,4 +1,4 @@
-package hous.server.config.interceptor;
+package hous.server.config.interceptor.auth;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
