+++ src/main/java/hous/server/config/interceptor/version/VersionInterceptor.java
@@ -1,11 +1,12 @@
-package hous.server.config.interceptor;
+package hous.server.config.interceptor.version;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.web.method.HandlerMethod;
 import org.springframework.web.servlet.HandlerInterceptor;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 
 @RequiredArgsConstructor
 @Component
@@ -15,7 +16,7 @@ public class VersionInterceptor implements HandlerInterceptor {
 
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response,
-            Object handler) throws Exception {
+                             Object handler) throws Exception {
         if (!(handler instanceof HandlerMethod)) {
             return true;
         }
