+++ src/main/java/hous/server/controller/rule/RuleController.java
@@ -2,7 +2,7 @@ package hous.server.controller.rule;
 
 import hous.server.common.dto.ErrorResponse;
 import hous.server.common.dto.SuccessResponse;
-import hous.server.config.interceptor.Auth;
+import hous.server.config.interceptor.auth.Auth;
 import hous.server.config.resolver.UserId;
 import hous.server.service.rule.RuleService;
 import hous.server.service.rule.dto.request.CreateRuleRequestDto;
